requirements.txt
discord.py
firebase-admin
apscheduler
openai
requests
python-dotenv
